requirements.txt
# 深度学习框架
torch>=1.7.0
torchvision>=0.8.0
torchaudio>=0.7.0
ultralytics>=8.0.0
onnxruntime>=1.15.0

# YOLOv7相关依赖
# YOLOv7使用原生PyTorch实现，无需额外的ultralytics包
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
seaborn>=0.11.0

# Web框架和API
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.7
requests>=2.25.0

# 图像处理
opencv-contrib-python>=4.5.0
scikit-image>=0.18.0

# 数据处理
pandas>=1.3.0
scipy>=1.7.0

# 可视化
plotly>=5.0.0
streamlit>=1.0.0

# 工具库
tqdm>=4.60.0
PyYAML>=5.4.0
tensorboard>=2.7.0

# 部署相关
gunicorn>=20.1.0
waitress>=2.0.0

# 开发工具
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0